Remove legacy E2E test harness
